refactor: 폴더명 Modal -> Overlay로 변경 #83
